pyproject.toml
[project]
name = "ace-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "openai>=2.6.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "temporalio>=1.9.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
]
